qcacld-3.0: Correct the tlv length of the hangdata
The Tlv of the hang data is reduced to 2bytes from the 4bytes
change the structure definitions accordingly.

CRs-Fixed: 2753804

+++ core/cds/src/cds_api.c
@@ -93,7 +93,7 @@ static struct __qdf_device g_qdf_ctx;
 static uint8_t cds_multicast_logging;
 
 struct cds_hang_event_fixed_param {
-	uint32_t tlv_header;
+	uint16_t tlv_header;
 	uint8_t recovery_reason;
 	char driver_version[11];
 	char hang_event_version;
